Add FSM changes
Change the GET_FSM so that if we start by assking HEAD requests (which
it will by default), the FSM will loop back on getting wnough responses
to the state requests.

If the HEAD requests indicate a conflict, then it will loop back to the
"prepare" stage, but this time do as a "legacy" GET request.  If there
is no conflict, then just the required object will be re-requested by
looping back to the "execute" stage but with a get_core that only
requires the one response.

There are still some big outstanding problems
- getting read_repair to work as we loop back aorund updating Results
- discarding a head_only request when we're expecting GETs