Preserve results for read repair
Change the logic to preserve the results for finalise operations.  When
a HEAD request has reached quorum, and there is a need for a follow-up
get this is now termed an UPDATE process.  The GET should update the
results, not start with a fresh set of results.  This means that after
merging to get the actual response, the full result set is available.

Also, this deals with the lated HEAD response problem.  If n is 3 and r
is 2, and we do an UPDATE, normally it would be expected that the thirs
HEAD response would arrive before the response to the updating GET.  How
should this be dealt with?  Discarding the HEAD response would give
equivalent behaviour to current Riak?  However, there may be valuable
information in the response (i.e. that a vnode needs read repair, or
that a vnode has a sibling, or even that a vnode has a superior answer).

So the slow(er) vnode response will form part of the re-merging to find
the result once the GET is received.  This may have some down-sides
(i.e. if there is a slow node, it is now more likely to be pressured by
read repair, and so there could be a spiral affect?).

Lat error response will just be added to the result set, and then
filtered out when head_merge or merge is called.  Think this is OK?