Unit tests added
This changes the behaviour significantly.

There are three types of requests that the get_fsm can make
- GET
- HEAD
- UPDATE

The GET requests is used only if that has been set as the default.  The
behaviour under GET should be the same as prior to this branch.

The HEAD request is used as the initial request if HEAD is set to the
default method.  The get core record is changed to indicate that this
will need a head_merge not a merge.  The HEAD request may receive GET
responses if there is a backend that does not have the HEAD capability
within the cluster.

When enough responses have been received, the get_core response method
will now call head_merge.  This will split the replies it can’t discard
into {Objects, HeadsToFetch} - where Objects are objects which are
ready to merge, and HeadsToFetch are head responses that now need to be
updated.

Normally there will be at this stage an empty Object list and a single
HeadtoFetch.  If there are siblings there may be multiple heads to
fetch.

The will prompt a request of type UPDATE - this will send a GET request
toe each of the nodes in the HeadsToFetch, and wait for a non-error
response from them all.  It will not increment the counter if it
receives a response from another node (although it will add it to the
results list).

When quorum is reached second time round - it is expected that now
there will be a set of Objects to merge, and no more HeadstoFetch.

This should deal with the problem of divorced head responses from the
original HEAD request disrupting the slow of UPDATE requests.  It also
simpler than before as there is not a distinct scenario fetch_all to be
used only when there are siblings.