HEAD and MD_CACHE
The implementation of the head request had the impact of nullifying the
MD_CACHE for PUTs, in that the fetch before the PUT was not populating
the cache on two out of three nodes.  the PUT was then requiring extra
GETs to complete.

This will now also update the cache for HEAD requests (as only header
information is required by the cache).