Add support for using HEAD requets in PUT path
If there is support for HEAD - then don't do a GET unless the vclock
indicates the need for a merge.  Reuses the logic set-aside for the
metadata cache to do a HEAD request in place of a cache fetch if the
cache is not defined (and otherwise will rely on the cache fetch).