Add snap outcome
Allow a snap_fold to be enabled, which should return an async folder where a snapshot has already been taken.

The idea is that snapshotted folders don't need to be concurrently to provide a close-to-consistent outcome - they can be handled in the core_node_worker pool not the core_vnode_worker_pool.