Clarify naming for snap_fold
Kept getting confused myself about naming.  Now the cpability is called snap_prefold (can a snapshot be made prior to a fold).

Bitcask backend is not yet tackled this way.  to iterate over bitcask, a new bitcask database needs to be opened (based on the current on-disk view of that database).  This open() process is expensive, and we don't want the vnode locked whilst we wait for the snapshot.

Proposal, not yet implemented is to extend the async options with which the backend can return to a fold_keys request.

Currently, either a response of:

{async, Folder} or

{queue, Folder} can be returned - where {queue, Folder} is only used if a snap_prefold was requested.

Idea is to be able to return:

{snap, ItrFun, FoldGenFun}

Where ItrFun will return snapshot:

Itr = ItrFun(),

and FoldGenFun will return a Folder when passed the snapshot:

Folder = FoldGenFun(Itr)

The idea then would be to extend the riak_kv_vnode/riak_core_vnode and the riak_kv_worker so that the core_vnode_worker_pool would be passed as the Work:

{snap, ItrFun, QueueFun, FoldGenFun, FinishFun}

(as opposed to {fold, FoldFun, FinishFun})

the riak_kv_worker would then (as part of the core_vnode_worker_pool) do:

Itr = ItrFun(),
FoldFun = FoldFunGen(Itr),
QueueFun({fold, FoldFun, FinishFun})

Where QueueFun would be a call to riak_core_node_worker_pool:handle_work.

So the snap would take place async in the vnode_worker_pool, but the fold would still be queued in the node_worker_pool.  So the output of the node_worker_pool would be at a roughly consistent point in the time, whilst all the fold work would still be throttled.