Add warning re multi and prefix_multi
There is limited testing of multi backend. Even having multi backend and only a single backend may have unexpected outcomes (for example even if the lone backend is leveled, Tictac AAE will be forced to run in parallel mode).

Multi backend has some hard problems, as it has to be open to other backends being added in the future.

There are potentially significant issues when combining muliple backends and active anti-entropy (especially if multi-backend is used to "drop" buckets).

No update has been made to 'prefix_multi' to adjust for the leveled backend and any new capabilities.  The 'prefix-multi' is now assumed to be unused outside of riak_cs