Nhse o34 orkv.i42 aae tree exclude (#43)
* Add support for bucket filter on aae_tree_exclude

* For full-sync - don't KeyFilter if Filter already present

If this is an attempt to full-sync a specific bucket or bucket-list - we should assume there is a specific intent to repair each bucket: so do not filter based on aae_tree_exclude.

* Update riak_pb

* Add SyncOnWrite to PUT properties on PB

Already included on HTTP

* Add summary documentation of bucket properties

* Information on changing properties

* Update following discussion of dual node failure

On two node failures, there is good reason to configure PR to at least 1, if notfound_ok is true.  Reminded of scenario in discussion with Adrian.

* Add utility function to reset key filter functions

Clears any cache, so that a new property is recognised on future alterations/additions.

* Renaming

To avoid confusion about inclusion or exclusion being true/false

* Update following review

* Update rebar.config following PR merges