Extract properties or options in bulk
Rather than repeatedly calling `lists:keyfind/2` or `proplists:get_value/3`

However, there are issues with the way bucket properties are merged with defaults.  Depending on the function used, thye could be merged using ukeysort - so there was a unique enty for every property, or keysort.  If it was merged with keysort there would be multiple properties, but only the first would be found (by proplists:get_value or lists:keyfind).

This now tries to standardise bucket property fetches (that may impact the PUT path) to produce genuinely merged properties.