Done't check siblings/size if not coordinating
The coordinator should detect this.  If it has been accepted elsewhere (i.e. it is a replicated PUT, or it has been added on a misconfigured node), then accept it here.

Changes the layout of the code to separate the size_check and sibling_check rather than split the encode_and_put fun.

This should reduce the number of calls to application:get_env (and consequent calls to ets:lookup/2)