Marginal gains
Walk the options on prepare (as well as validate) within the PUT_FSM

Fix the object_format to v1 (tests will need to intercept the function)

Only perform the size and sibling checks on a normal PUT at the coordinator.

Allow bucket properties to be kept in a map at the riak_kv_put_fsm, and still have options fetched.

Only calculate node_confirms, if node_confirms is a requirement for `enough`.  Calculate it more efficiently when required

Revert changes to remove put functions from riak_client - required for riak_test.