WIP: add vnode_object_cache to PUT flow
TODO: add to GET flow

On counter updates to backend, the read before the write is the biggest cost.  As the coordinator merges with an empty object every update - a GET not a HEAD is required.

counters are small, so caching recently updated counters should be inexpensive.  They can be cached in a non-binary form, saving on de-serialisation as well as fetch time.

Should only be cached for the specific buckets (by bucket property), and if a cache is enabled.

Cache is currently size in objects.