Nhse o34 leveled.i433 enhancequery (#51)
Add support for new query API.

The riak_kv_query module allows a query to be constructed with validation checks.

The riak_kv_query_server is a gen_Server for managing a query.  The coverage_fsm behaviour is not used, to aid future migration from the deprecated fsm.

There is a riak_kv_query_buffer added to replace the use of riak_kv_fold_buffer - which assumes the accumulator is a list.

* Update backend specs

Resolves OTP 24 dialyzer complaints after clarifying specs in backend behaviour

* Use OpenRiak meck

upstream meck has updated master, and no longer supports OTP < 25

* Improve stats

Have both vnode and node stats (as with other operations)

* Add WebMachine API for Query

Add HTTP API for Query - allowing for PSOT of JSON with JSON response.

* Ordering of keys/term_with_keys

The keys may be read in every oder - the keys are not ordered with the term.

Likewise with term_with_keys as the term may now be extracted from within the term (i.e. a projected attribute)

* Update client branch to track updates

* Allow for the API to pass a function to encode the results

Ensures that the message from the query server back to the client is a binary reference pass not a copy.

* Add raw_keys support and improve performance

The de-duplication of keys using lists:umerge/2 does not scale effectively,  Now, beyond a limit a temporary ets table is used instead for de-duplication.

Still, the performance was slower than current 2i for large key sets > 10K.  This is as riak_kv_query_server is the bottleneck - it cannot ACK batches of keys whilst it is handling previous batches.  In contrast the existing 2i simply forwards batches immediately - so that any delay processing is on the riak_client process queue, not on the query coordinator.

To help with this performance issue, the default batch size is increased, and also jittered to help avoid all vnode workers sending concurrently.

The use of an ets table requires every key to be wrapped in a tuple.  This is done at source (in the buffer), to avoid a further delay on the query server running a lists:map/2 function.  The JSON encoder will handle the tuple wrapping of the key (although it could be efficiently removed in the ets case using a select statement rather than tab2list.

* Add acceleration for sorted term_with_keys

Both keys and term_with_keys are now sorted and de-duplicated, and using a consistent code path.

* Add testing for invalid filter_query

* Start query stat monitor

* Change names of accumulation options

* Add support for raw_terms

* Resolve issues with combo queries

* Add support for max_results and continuation

* Revert moving of aae_fold types

Add unnecessary noise to change

* Minor speed-up in JSON

Using the character not the string speeds up iolist_to_binary

* Initial commit of doc - WIP

* Update QueryAPI.md

Play around to resolve markdown parsing issue

* Update Query API - WIP

* Markdown wrangling

* Markdown wrangling

* Update Query API - WIP

* Update Query API - WIP

* Add json encoding tags to markdown snippets

* Substitutions not a list

* Update Query API - corrections following tests

* Update QueryAPI.md

Resolve broken headings

* Remove MAPFOLD - prototype replaced by Query API

* Add simplified example first

As suggested in review.

TODO left to finish performance section.

* Use correct branches

Updated RHC is not required internally within riak_kv - just by riak_test.

* Formatting in doc

* Doc changes following review

* Further updates post review

* Switch to iolist_to_binary/1 - more forgiving

* Don't support timeout refresh

No infrastructure to trigger it, and it wouldn't have worked.

* More review feedback