Convert riak object metadata into a map (#65)
* Convert riak object metadata into a map

The map module is mroe performant that the dict module, in part as dict is coupled to an old (deprecated) hash algorithm.

All objects are stored in binary form - which is independent of the data type.  This code will still handle objects converting metadata to dict.  However, enabling the switch to version 1 will require all connected nodes to support metadata as map.

A capability could be used, but there is history of capability registrations taking a long time to remove - so the shortcut of making this an operator responsibility has been made, especially with a major version change pending.

* Revert unrelated API change from this PR

* Use map to process headers and strip/re-add x-riak-meta- prefix

The PB API will does not have x-riak-api- metadata prefix on user metadata keys - and so the riak_kv_wm_object cannot assume its there.

* Fix conditional PUT issue (introduced by previous commit)

* Revert branch to openriak-3.4

* Revert rebar change - riakc a dep of rhc

changes to riakc now merged in