Nhse o34 orkv.i84 querygrant (#85)
* Add configurable support for exempting ops requests for authorisation

Default will change, to enforce authorisation on all requests when security i s enabled

* Add security section to Ops guide

* Slight doc change

* Apply suggestions from code review

Co-authored-by: A. G. Madi <WarpEngineer@users.noreply.github.com>

* Update to security source doc following review

Added a caveat to indicate that user/password are not really well named.  There's basic protection of passwords through hashing etc, but probably not the protection that might be expected for managing the login credentials of real people.

---------

Co-authored-by: A. G. Madi <WarpEngineer@users.noreply.github.com>