Add delete_mode to schema, with some tests of "invalid" values (#72)
* Add delete_mode to schema, with some tests of "invalid values"

In theory any positive integer is possible for timeout.  But deleting 1000s a second would require a lot of timer process to be spawned if the limit was very high?

Limiting to some (hopefully) sane limit - but allowing it to be overridden in advanced.config (as with compromise on configuration of large ring_size).

* Add doc changes for setting delete mode